% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract}
\alias{extract}
\title{extract particular values from a retrieved set of RfC metadata}
\usage{
extract(rfcs, field)
}
\description{
foo bar
}
\examples{
rfc_results <- lapply(c(30,40,50), retrieve_rfc)
authors <- extract(rfc_results, "authors")
}
\seealso{
\code{\link{retrieve_rfc}} for retrieving the data.
}

