% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFc.R
\name{Release notes}
\alias{Release notes}
\title{Release notes}
\description{
\strong{release 0.1-2}\cr
Support for obtaining available datasets and varaibles with\cr
     fcDataSets()\cr
     fcVaraibles()\cr
     \cr
\strong{release 0.1-1}\cr
Support for fetching simple time series and grids
}

