% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retrieve_rfc}
\alias{retrieve_rfc}
\title{grab an IETF RfC and extract metadata}
\usage{
retrieve_rfc(id)
}
\arguments{
\item{id}{an RfC ID (e.g. "2000")}
}
\description{
\code{\link{retrieve_rfc}} lets you get an IETF RfC and extract the associated
metadata, including its authors, status, title, abstract, date of publication and what
RfCs it rendered obsolete (if any)
}
\examples{
results <- retrieve_rfc(1138)
}

