\name{RFc-package}
\alias{RFc-package}
\alias{RFc}
\docType{package}
\title{
R client for FetchClimate service
}
\description{
Extracts environmental data (such as air temperature, precipitation rate, wind speed, etc.) from FetchClimate WEB service\cr
for the specified geo-locations and time bounds.
}
\details{
\tabular{ll}{
Package: \tab RFc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-03-25\cr
License: \tab MIT\cr
}
RFc is realy great package!!!\cr
functions overview

}
\author{
Dmitry A. Grechka

Maintainer: Dmitry A. Grechka <dmitryg@itis.cs.msu.ru>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
