% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFc.R
\docType{package}
\name{RFc}
\alias{RFc}
\alias{RFc-package}
\title{RFc: R client for FetchClimate service}
\description{
Extracts raw, averaged environmental data (such as air temperature, precipitation rate, wind speed,
etc.) published at the FetchClimate Web service (<http://fc.itis.cs.msu.ru/>) for the specified geo-locations and time bounds from
different data sets.
}
\section{Time series fetching functions}{

fcTimeSeriesYearly, fcTimeSeriesDaily, fcTimeSeriesHourly
}

\section{Gridded data fetching functions}{

fcGrid
}

